dags/requirements.txt
apache-airflow==1.10.14
apache-airflow[kubernetes]==1.10.14
kubernetes
PyYAML
